beginnings of handlers for more soz api fns
[git-p4: depot-paths = "//depot/Tumri/tas/joz/": change = 6508]

+++ src/com/tumri/joz/jozMain/MappingObj.java
@@ -8,7 +8,7 @@ package com.tumri.joz.jozMain;
 
 public class MappingObj
 {
-    public MappingObj (MappingType type, float weight, int modified,
+    public MappingObj (MappingType type, float weight, long modified,
 		       String t_spec)
     {
 	_type = type;
@@ -24,13 +24,13 @@ public class MappingObj
 
     public MappingType get_type () { return _type; }
     public float get_weight () { return _weight; }
-    public int get_modified () { return _modified; }
+    public long get_modified () { return _modified; }
     public String get_t_spec () { return _t_spec; }
 
     // implementation details -------------------------------------------------
 
     private MappingType _type;
     private float _weight;
-    private int _modified; // in seconds since jan 1, 1970
+    private long _modified; // in seconds since jan 1, 1970
     private String _t_spec;
 }
